==> code/1b_salmon_gadidae_taxonomic_assignment.Rmd <==
---
title: "oncorhynchus_gadid_customdbs"
author: "Kimberly Ledger"
date: "2024-07-08"
output: html_document
---

```{r load-libraries}
library(tidyverse)
rename <- dplyr::rename
```

input blastn results from custom salmon and gadid databases
```{r}
oncor <- read.table("/home/kimberly.ledger/dbo_metabarcoding/outputs/oncorhynchus_results.txt")  %>%
  filter(V4 > 100) %>% # make sure all retained matches are >100 bp
  select(-V5, -V6, -V7, -V8, -V9, -V10, -V11, -V12, -V13, -V14) %>%  #remove unnecessary columns
  group_by(V1, V2) %>% # group by the sequence key and the full taxonomy to reduce duplicate entries
  unique() %>% # doing that reduced the number of entries
  rename(qseqid=V1, species=V2, perc_id=V3, length=V4) %>% #rename headers
  filter(perc_id > 98)
```

## how many salmon asvs
```{r how-many-total-ASVs-w-taxonomy}
oncor %>% 
  ungroup() %>%
  select(qseqid) %>%
  unique() %>%
  nrow
```

## rank the hits by % identity
```{r}
seq_id_diff <- oncor %>%
  select(-length) %>%
  group_by(qseqid, species) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
      #mutate(id_rank = rank(desc(seq_percID), ties.method = "min")) %>% # rank the taxonomic hits per sequence by % id
       mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
      arrange(diff)

seq_id_diff
```

since there are very few basepair difference among species, let's go with species within 0.5% of the top match 
```{r}
keepers <- seq_id_diff %>%
  filter(diff < 0.5)
```


## so now determine whether ASVs should be assigned to genus, family, order, etc. 
```{r}
singletons <- keepers %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
singleton_df <- singletons %>%
  left_join(keepers) %>%
  select(-n) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species) #%>%
  #select(qseqid, taxon, taxonomic_level)

## Genus-level matches
# remove the singletons from the bigger df 
single_genus <- keepers %>%
  anti_join(singleton_df)%>% # at best, these should be genus-level matches
  left_join(., keepers) %>%
  arrange(qseqid) %>%
  mutate(taxonomic_level = "genus") %>%
  mutate(taxon = "Oncorhynchus") #%>%
  #select(qseqid, taxon, taxonomic_level) %>%
  #unique
```

so as I have the Oncorhynchus genus designation right now, it can be any pacific salmon (or O. mykiss) except for Chinook. 

remove underscore from taxon name 
```{r}
oncorhynchus_df <- singleton_df %>%
  bind_rows(single_genus)

oncorhynchus_df$taxon <- gsub("_", " ", oncorhynchus_df$taxon)
oncorhynchus_df$species <- gsub("_", " ", oncorhynchus_df$species)
oncorhynchus_df$genus <- "Oncorhynchus"
oncorhynchus_df$family <- "Salmonidae"
oncorhynchus_df$order <- "Salmoniformes"
oncorhynchus_df$class <- "Actinopteri"
oncorhynchus_df$phylum <- "Chordata"
oncorhynchus_df$kingdom <- "Eukaryota"

#export 
write.csv(oncorhynchus_df, "../outputs/oncorhynchus_asv_id.csv")
```


now repeat for gadids 
```{r}
gadidae <- read.table("/home/kimberly.ledger/dbo_metabarcoding/outputs/gadidae_results.txt")  %>%
  filter(V4 > 100) %>% # make sure all retained matches are >100 bp
  select(-V5, -V6, -V7, -V8, -V9, -V10, -V11, -V12, -V13, -V14) %>%  #remove unnecessary columns
  group_by(V1, V2) %>% # group by the sequence key and the full taxonomy to reduce duplicate entries
  unique() %>% # doing that reduced the number of entries
  rename(qseqid=V1, species=V2, perc_id=V3, length=V4) %>% #rename headers
  filter(perc_id > 98)
```

## how many gadidae asvs
```{r how-many-total-ASVs-w-taxonomy}
gadidae %>% 
  ungroup() %>%
  select(qseqid) %>%
  unique() %>%
  nrow
```

## rank the hits by % identity
```{r}
seq_id_diff <- gadidae %>%
  select(-length) %>%
  group_by(qseqid, species) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
      #mutate(id_rank = rank(desc(seq_percID), ties.method = "min")) %>% # rank the taxonomic hits per sequence by % id
       mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
      arrange(diff)

seq_id_diff
```

since there are very few basepair difference among species, let's go with species within 0.5% of the top match 
```{r}
keepers <- seq_id_diff %>%
  filter(diff < 0.5)
```


## so now determine whether ASVs should be assigned to genus, family, order, etc. 
```{r}
singletons <- keepers %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
singleton_df <- singletons %>%
  left_join(keepers) %>%
  select(-n) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species) #%>%
  #select(qseqid, taxon, taxonomic_level)

# if the match is Gadus macrocephalus and Gadus chalcogrammus - then go with the genus level 

## Genus-level matches
# remove the singletons from the bigger df 
single_genus <- keepers %>%
  anti_join(singleton_df)%>% # 
  left_join(., keepers) %>%
  separate(col = species, into = c("Genus", "spp"), remove = F) %>%
  select(qseqid, Genus) %>%
  unique() %>%
  tally() %>%
  filter(n == 1)

gadus <- keepers %>%
  anti_join(singleton_df)%>% # 
  left_join(., keepers) %>% 
  filter(qseqid %in% single_genus$qseqid) %>%
  #select(qseqid, species) %>% 
  #unique() %>% 
  mutate(taxon = "Gadus") %>%
  mutate(taxonomic_level = "genus")
  
## what's remaining? 
leftover <- keepers %>%
  anti_join(singleton_df)%>% # 
  left_join(., keepers) %>%
  filter(!qseqid %in% single_genus$qseqid) %>%
  #select(qseqid, species) %>%
  #unique() %>%
  mutate(taxon = "Gadidae") %>%
  mutate(taxonomic_level = "family")
```

remove underscore from taxon name 
```{r}
gadidae_df <- singleton_df %>%
  bind_rows(gadus) %>%
  bind_rows(leftover)

gadidae_df$taxon <- gsub("_", " ", gadidae_df$taxon)
gadidae_df$species <- gsub("_", " ", gadidae_df$species)
gadidae_df <- gadidae_df %>%
  mutate(genus = ifelse(species == "Gadus chalcogrammus", "Gadus", NA),
         genus = ifelse(species == "Gadus macrocephalus", "Gadus", genus),
         genus = ifelse(species == "Boreogadus saida", "Boreogadus", genus),
         genus = ifelse(species == "Eleginus gracilis", "Eleginus", genus),
         genus = ifelse(species == "Microgadus proximus", "Microgadus", genus),
         genus = ifelse(species == "Arctogadus glacialis", "Arctogadus", genus))
gadidae_df$family <- "Gadidae"
gadidae_df$order <- "Gadiformes"
gadidae_df$class <- "Actinopteri"
gadidae_df$phylum <- "Chordata"
gadidae_df$kingdom <- "Eukaryota"

#export 
write.csv(gadidae_df, "/home/kimberly.ledger/dbo_metabarcoding/outputs/gadidae_asv_id.csv")
```
